+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+version 1.6.0
+-------------
++ Faster runtime due to faster gene identification, sequence overview creation 
+  and IMGT TAR archive creation.
++ Two extra IMGT files are provided. One for IGM with less than 2% mutations
+  (naive) and one for IGM with more than 2% mutations (naive memory).
++ All IMGT files per sequence class are always provided regardless of how the 
+  ``Fast`` option is set. Previously this had to be set to ``no``.
++ Fix a bug in sequence overview where links to unmatched sequences where not
+  working properly in the by_id.html file.
+
 version 1.5.0
 -------------
 + Add an option to download all output files in a zip file.